Feat. Step 1.1 프로젝트 설정 완료
Why:
- Phase 1 MVP 개발을 위한 기본 프로젝트 구조 설정
- MVVM 아키텍처 적용 준비
- App Sandbox 권한 설정

How:
- MVVM 폴더 구조 생성 (Models, Views, ViewModels, Services)
- Info.plist 추가 (카메라/마이크 권한 설명)
- MacDisplayLink.entitlements 추가 (App Sandbox 설정)
- 기존 파일 삭제 및 임시 ContentView 생성
- 빌드 테스트 성공 확인
