--- MacDisplayLink/CaptureDeviceManager.swift
@@ -1,73 +1,0 @@
-//
-//  CaptureDeviceManager.swift
-//  MacDisplayLink
-//
-//  Created by Codex.
-//
-
-import AVFoundation
-import Combine
-import Foundation
-
-final class CaptureDeviceManager: ObservableObject {
-    @Published private(set) var videoDevices: [AVCaptureDevice] = []
-    @Published private(set) var audioDevices: [AVCaptureDevice] = []
-    private var cancellables = Set<AnyCancellable>()
-
-    init() {
-        refreshVideoDevices()
-        refreshAudioDevices()
-        startObservingDeviceChanges()
-    }
-
-    func refreshVideoDevices() {
-        let discoverySession = AVCaptureDevice.DiscoverySession(
-            deviceTypes: [.external],
-            mediaType: .video,
-            position: .unspecified
-        )
-        videoDevices = discoverySession.devices.filter { isExternalCaptureDevice($0) && !isBuiltInCamera($0) }
-    }
-
-    private func isBuiltInCamera(_ device: AVCaptureDevice) -> Bool {
-        #if os(macOS)
-        return device.deviceType == .builtInWideAngleCamera
-        #else
-        switch device.deviceType {
-        case .builtInWideAngleCamera,
-             .builtInUltraWideCamera,
-             .builtInTelephotoCamera,
-             .builtInDualCamera,
-             .builtInDualWideCamera,
-             .builtInTripleCamera,
-             .builtInTrueDepthCamera:
-            return true
-        default:
-            return false
-        }
-        #endif
-    }
-
-    private func isExternalCaptureDevice(_ device: AVCaptureDevice) -> Bool {
-        device.deviceType == .external
-    }
-
-    private func startObservingDeviceChanges() {
-        let center = NotificationCenter.default
-        let connected = center.publisher(for: AVCaptureDevice.wasConnectedNotification)
-        let disconnected = center.publisher(for: AVCaptureDevice.wasDisconnectedNotification)
-
-        connected
-            .merge(with: disconnected)
-            .receive(on: DispatchQueue.main)
-            .sink { [weak self] _ in
-                self?.refreshVideoDevices()
-                self?.refreshAudioDevices()
-            }
-            .store(in: &cancellables)
-    }
-
-    private func refreshAudioDevices() {
-        audioDevices = AVCaptureDevice.devices(for: .audio)
-    }
-}
