+++ MacDisplayLink/Info.plist
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
+<plist version="1.0">
+<dict>
+	<key>NSCameraUsageDescription</key>
+	<string>캡쳐 카드의 비디오 입력을 받기 위해 카메라 권한이 필요합니다.</string>
+	<key>NSMicrophoneUsageDescription</key>
+	<string>캡쳐 카드의 오디오 입력을 받기 위해 마이크 권한이 필요합니다.</string>
+</dict>
+</plist>
